+++ pandapower/test/control/test_discrete_tap_control.py
@@ -288,6 +288,77 @@ def test_discrete_tap_control_hv_from_tap_step_percent():
     assert net.trafo.tap_pos.at[0] == -1
 
 
+def test_discrete_tap_control_hv_from_tap_step_percent__negative_tap_step():
+    # --- load system and run power flow
+    net = simple_four_bus_system()
+    set_user_pf_options(net, init='dc', calculate_voltage_angles=True)
+    # --- initial tap data
+    net.trafo.tap_side = 'hv'
+    net.trafo.tap_neutral = 0
+    net.trafo.tap_min = -2
+    net.trafo.tap_max = 2
+    net.trafo.tap_step_percent = -1.25
+    net.trafo.tap_pos = 0
+    # --- run loadflow
+    runpp(net)
+
+    DiscreteTapControl.from_tap_step_percent(net, 0, side='lv', vm_set_pu=0.98)
+
+    logger.info("case1: low voltage")
+    logger.info("before control: trafo voltage at low voltage bus is %f, tap position is %u"
+                % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+
+    # run control
+    runpp(net, run_control=True)
+    logger.info(
+        "after DiscreteTapControl: trafo voltage at low voltage bus is %f, tap position is %f"
+        % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+    assert net.trafo.tap_pos.at[0] ==  -1
+
+    # check if it changes the lower and upper limits
+    net.controller.object.at[0].vm_set_pu = 1
+    runpp(net, run_control=True)
+    assert abs(net.controller.object.at[0].vm_upper_pu - 1.00725) < 1e-6
+    assert abs(net.controller.object.at[0].vm_lower_pu - 0.99275) < 1e-6
+    net.controller.object.at[0].vm_set_pu = 0.98
+    runpp(net, run_control=True)
+    assert abs(net.controller.object.at[0].vm_upper_pu - 0.98725) < 1e-6
+    assert abs(net.controller.object.at[0].vm_lower_pu - 0.97275) < 1e-6
+
+    # increase voltage from 1.0 pu to 1.03 pu
+    net.ext_grid.vm_pu = 1.03
+    # switch back tap position
+    net.trafo.at[0, "tap_pos"] = 0
+    runpp(net)
+
+    logger.info("case2: high voltage")
+    logger.info("before control: trafo voltage at low voltage bus is %f, tap position is %u"
+                % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+
+    # run control
+    runpp(net, run_control=True)
+    logger.info(
+        "after DiscreteTapControl: trafo voltage at low voltage bus is %f, tap position is %f"
+        % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+    assert net.trafo.tap_pos.at[0] == -2
+    # reduce voltage from 1.03 pu to 0.969 pu
+    net.ext_grid.vm_pu = 0.969
+    # switch back tap position
+    net.trafo.at[0, "tap_pos"] = 0
+    runpp(net)
+
+    logger.info("case2: high voltage")
+    logger.info("before control: trafo voltage at low voltage bus is %f, tap position is %u"
+                % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+
+    # run control
+    runpp(net, run_control=True)
+    logger.info(
+        "after DiscreteTapControl: trafo voltage at low voltage bus is %f, tap position is %f"
+        % (net.res_bus.vm_pu[net.trafo.lv_bus].values.item(), net.trafo.tap_pos.values.item()))
+    assert net.trafo.tap_pos.at[0] == 1
+
+
 def test_discrete_tap_control_vectorized_lv():
     # --- load system and run power flow
     net = create_empty_network()