+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Change Log
 
 [upcoming release] - 2026-..-..
 -------------------------------
+- [FIXED] `DiscreteTapControl` to work with `negative tap_step_percent`.
 - [CHANGED] updated the contributing file and documentation
 - [ADDED] `allow_duplicate_index` parameter to `reindex_buses` with default to `false`.
 - [ADDED] `add_basic_std_types` parameter to `from_excel`